requirements.txt
Jinja2==2.7.3
PyYAML==3.11
docopt==0.6.1
pykwalify==14.08
python-etcd==0.3.2
